moving critics to its own directory similar to src as pre-established
